refactored multilayerconfig to accomodate API changes. Add layer method in ListBuilder. Test to be written. Build Success
